+++ src/lib/drag-drop.ts
@@ -1,13 +1,15 @@
-type DragPlaylist = {
+type DraggedTracks = {
+  ids: string[]
+  playlistIndexes?: number[]
+}
+type DraggedPlaylist = {
   id: string
   fromFolder: string
   level: number
 }
 
-export const dragged = {
-  tracks: {
-    indexes: [] as number[],
-    ids: [] as string[],
-  },
-  playlist: null as null | DragPlaylist,
+type Dragged = {
+  tracks?: DraggedTracks
+  playlist?: DraggedPlaylist
 }
+export const dragged: Dragged = {}